requirements_updated.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
supabase>=2.0.0
postgrest>=0.13.0

pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0

python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

scikit-learn>=1.3.0
joblib>=1.3.0

spacy>=3.6.0
nltk>=3.8.0
